Move privileges test into main test file, test Helm 2 && 3
