+++ install/test/helm_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/solo-io/gloo/projects/gloo/pkg/plugins/wasm"
 	"github.com/solo-io/gloo/test/matchers"
 	"github.com/solo-io/go-utils/installutils/kuberesource"
+	"github.com/solo-io/go-utils/manifesttestutils"
 	"k8s.io/apimachinery/pkg/util/sets"
 
 	"github.com/gogo/protobuf/proto"
@@ -2248,6 +2249,36 @@ metadata:
 					testManifest.ExpectDeploymentAppsV1(ingressDeploymentPostMerge)
 				})
 			})
+
+			Describe("Deployment Privileges Test", func() {
+				Context("Gloo", func() {
+					Context("all cluster-scoped deployments", func() {
+						It("is running all deployments with non root user permissions by default", func() {
+
+							prepareMakefile(namespace, helmValues{})
+
+							expectNonRoot(testManifest)
+						})
+
+						It("is running all deployments with non root user permissions with knative, accessLogger, ingress, and mTLS enabled", func() {
+
+							prepareMakefile(namespace, helmValues{
+								valuesArgs: []string{
+									"gateway.enabled=false",
+									"settings.integrations.knative.enabled=true",
+									"settings.integrations.knative.version=v0.10.0",
+									"accessLogger.enabled=true",
+									"ingress.enabled=true",
+									"global.glooMtls.enabled=true",
+								},
+							})
+
+							expectNonRoot(testManifest)
+						})
+					})
+				})
+
+			})
 		})
 	}
 
@@ -2261,4 +2292,42 @@ func cloneMap(input map[string]string) map[string]string {
 	}
 
 	return ret
+}
+
+func expectNonRoot(testManifest manifesttestutils.TestManifest) {
+	deployments := testManifest.SelectResources(func(resource *unstructured.Unstructured) bool {
+		return resource.GetKind() == "Deployment"
+	})
+
+	Expect(deployments.NumResources()).NotTo(BeZero())
+
+	deployments.ExpectAll(func(resource *unstructured.Unstructured) {
+		rawDeploy, err := resource.MarshalJSON()
+		Expect(err).NotTo(HaveOccurred())
+
+		deploy := appsv1.Deployment{}
+		err = json.Unmarshal(rawDeploy, &deploy)
+		Expect(err).NotTo(HaveOccurred())
+
+		Expect(deploy.Spec.Template).NotTo(BeNil())
+
+		podLevelSecurity := false
+		// Check for root at the pod level
+		if deploy.Spec.Template.Spec.SecurityContext != nil {
+			Expect(deploy.Spec.Template.Spec.SecurityContext.RunAsUser).NotTo(Equal(0))
+			podLevelSecurity = true
+		}
+
+		// Check for root at the container level
+		for _, container := range deploy.Spec.Template.Spec.Containers {
+			if !podLevelSecurity {
+				// If pod level security is not set, containers need to explicitly not be run as root
+				Expect(container.SecurityContext).NotTo(BeNil())
+				Expect(container.SecurityContext.RunAsUser).NotTo(Equal(0))
+			} else if container.SecurityContext != nil {
+				// If podLevel security is set to non-root, make sure containers don't override it:
+				Expect(container.SecurityContext.RunAsUser).NotTo(Equal(0))
+			}
+		}
+	})
 }
